Added MenuBase.
And use it as base class for both `Menu` and `ContextMenu`.
